refactor: split transaction ingestion dtos
